requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
addict=2.4.0=pypi_0
blas=1.0=mkl
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.9.14=py38h06a4308_0
charset-normalizer=2.1.1=pypi_0
cycler=0.11.0=pypi_0
einops=0.4.1=pypi_0
fftw=3.3.9=h27cfd23_1
flopth=0.1.2=pypi_0
fonttools=4.37.1=pypi_0
fvcore=0.1.5.post20220512=pypi_0
google-auth=2.11.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.48.1=pypi_0
idna=3.3=pypi_0
imageio=2.21.2=pypi_0
importlib-metadata=4.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.10=pypi_0
joblib=1.1.0=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lpips=0.1.4=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mmcv=1.6.1=pypi_0
mmcv-full=1.6.2=pypi_0
monai=0.9.1=pypi_0
munch=2.5.0=pyhd3eb1b0_0
ncurses=6.3=h5eee18b_3
networkx=2.8.6=pypi_0
nibabel=4.0.2=pypi_0
numpy=1.23.2=pypi_0
oauthlib=3.2.0=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.4.4=pypi_0
pillow=9.2.0=pypi_0
pip=22.1.2=py38h06a4308_0
portalocker=2.5.1=pypi_0
protobuf=3.19.4=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.9=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytz=2022.2.1=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.4.1=pypi_0
setuptools=63.4.1=py38h06a4308_0
simpleitk=2.2.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.2=h5082296_0
tabulate=0.8.10=pypi_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.8.12=pypi_0
timm=0.6.7=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.12.1=pypi_0
torchvision=0.13.1=pypi_0
tqdm=4.64.1=pypi_0
typing-extensions=4.3.0=pypi_0
urllib3=1.26.12=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yacs=0.1.8=pypi_0
yapf=0.32.0=pypi_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h5eee18b_3
